+++ src/scenes/BirdScene.ts
@@ -79,8 +79,8 @@ export default class BirdScene extends Phaser.Scene {
       this.scene.pause();
       this.scene.launch('PauseScene', { scene: this.scene.key });
     });
-    InputManager.setLegacyGamepadKeyboardBridge(true);
-    this.events.once('shutdown', () => InputManager.setLegacyGamepadKeyboardBridge(false));
+    InputManager.setLegacyGamepadKeyboardBridge(true, this.scene.key);
+    this.events.once('shutdown', () => InputManager.setLegacyGamepadKeyboardBridge(false, this.scene.key));
 
     // Collisions
     this.physics.add.collider(this.bird, this.pipes, () => this.endGame());